Move GitLab common
