+++ destinations/airbyte-faros-destination/src/converters/gitlab-ci/groups.ts
@@ -1,7 +1,7 @@
 import {AirbyteRecord} from 'faros-airbyte-cdk';
 
-import {GitlabCommon, GitlabConverter} from '../common/gitlab';
 import {DestinationModel, DestinationRecord, StreamContext} from '../converter';
+import {GitlabCommon, GitlabConverter} from '../gitlab/common';
 
 export class Groups extends GitlabConverter {
   source = 'GitLab-CI';
