+++ destinations/airbyte-faros-destination/src/converters/gitlab/group_labels.ts
@@ -1,7 +1,7 @@
 import {AirbyteRecord} from 'faros-airbyte-cdk';
 
-import {GitlabConverter} from '../common/gitlab';
 import {DestinationModel, DestinationRecord, StreamContext} from '../converter';
+import {GitlabConverter} from './common';
 
 export class GroupLabels extends GitlabConverter {
   readonly destinationModels: ReadonlyArray<DestinationModel> = ['tms_Label'];
