+++ destinations/airbyte-faros-destination/src/converters/gitlab/jobs.ts
@@ -1,13 +1,13 @@
 import {AirbyteLogger, AirbyteRecord} from 'faros-airbyte-cdk';
 import {Utils} from 'faros-feeds-sdk';
 
-import {CategoryRef, GitlabCommon, GitlabConverter} from '../common/gitlab';
 import {
   DestinationModel,
   DestinationRecord,
   StreamContext,
   StreamName,
 } from '../converter';
+import {CategoryRef, GitlabCommon, GitlabConverter} from './common';
 
 export class Jobs extends GitlabConverter {
   private readonly logger: AirbyteLogger = new AirbyteLogger();
