+++ destinations/airbyte-faros-destination/src/converters/gitlab/merge_requests.ts
@@ -1,13 +1,13 @@
 import {AirbyteRecord} from 'faros-airbyte-cdk';
 import {Utils} from 'faros-feeds-sdk';
 
-import {GitlabCommon, GitlabConverter} from '../common/gitlab';
 import {
   DestinationModel,
   DestinationRecord,
   StreamContext,
   StreamName,
 } from '../converter';
+import {GitlabCommon, GitlabConverter} from './common';
 
 export class MergeRequests extends GitlabConverter {
   readonly destinationModels: ReadonlyArray<DestinationModel> = [
